ProjectFile/requirements.txt
streamlit
pandas
matplotlib
langchain
google-generativeai